Create pathfinder.py
This a Path finder program to find the shortest path (Manhattan distance) in between two given points by avoiding barriers. The basics of this technique can be used to develop navigation apps where we don't have to draw the barriers but instead have to identify them and give the shortest path between to points.